Restructure app with static components to fix flickering
