+++ src/types.ts
@@ -1,0 +1,5 @@
+export interface Todo {
+  id: string;
+  text: string;
+  completed: boolean;
+}
